requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
psycopg2-binary>=2.9.9
asyncpg>=0.28.0
alembic>=1.12.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
openai>=1.3.0
websockets>=11.0.3
bcrypt>=4.0.1
email-validator>=2.0.0
httpx>=0.25.0 